+++ CHANGELOG.md
@@ -72,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - **Breaking**: Change `ZfpExpertParams` to `ZfpExpertConfiguration` replacing existing `ZfpExpertConfiguration`
    - The previous `ZfpExpertConfiguration` was incorrect and was identical to `ZfpFixedRateConfiguration`
+ - Fix `FilesystemStore::list_prefix` with an empty prefix
 
 ## [0.14.0] - 2024-05-16